Get local species list

Using longitude and latitude coordinates and a buffer radius, this function retrieves a plant species list from GBIF and formats it for use in BotanizeR_quiz() or the BotanizeR Shiny app. If a backbone_list is supplied, it will be subset for the species found in GBIF and returned.

BotanizeR_getlocallist(
  long = NA,
  lat = NA,
  radius = 0.1,
  taxonKey = 7707728,
  limit = 10000,
  backbone_list = NA
)

Arguments

long
a numeric, defining the longitude of the wanted location in decimal degrees.

lat
a numeric, defining the latitude of the wanted location in decimal degrees.

radius
a numeric, defining the radius of the buffer around the wanted location in degrees. lat - radius needs to be > -90 and lat + radius needs to be < 90.

taxonKey
a numeric, defining the taxonKey of the target taxonomic group used by GBIF passed on to rgbif::occ_data. Default of 7707728 represents Tracheophyta. Use rgbif::name_lookup to find taxonKey of alternative taxa.

limit
a numeric, defining the maximum number of occurrences retrieved from GBIF passed on to rgbif::occ_data. Increasing it may significantly reduce speed.

backbone_list
a data.frame with a SPECIES column containing the species to be subset. If this list shall be used as species_list in BotanizeR_quiz(), It needs to contain at least the following columns: NAMNR, TAXONNAME, SPECIES and GENUS. See BotanizeR_quiz() for more info.

Value

A data.frame of the same structure like backbone_list including only those species found in GBIF for the given coordinates and radius. An additional column FREQis added providing the frequency of each for selecting more common species. If backbone_list is NA the species list obtained from GBIF is formatted to contain all essential columns to run BotanizeR_quiz() or to use it in the BotanizeR Shiny app.

References

Weigelt, P., Denelle, P., Brambach, F. & Kreft, H. (2021) A flexible
R-package with Shiny-App for practicing plant identification in times of
online teaching and beyond. submitted.

Examples

# Example

# Species list for Germany with IDs from floraweb.de
data(BotanizeR_species)

# Load species list for Britain and Germany with species and IDs from 
# https://www.floraweb.de and https://www.brc.ac.uk/plantatlas/
data(BotanizeR_species)

# Subset the species list for species occurring in Cairngorms National Park
BotanizeR_Cairngorms <- 
  BotanizeR_getlocallist(long = -3.5966, lat = 57.0877, 
                         radius = 0.5, taxonKey = 7707728, 
                         limit = 10000, backbone_list = BotanizeR_species)

# Create a new species list for the Island of Bornholm, Denmark
BotanizeR_Bornholm <- 
  BotanizeR_getlocallist(long = 14.9184, lat = 55.1273, 
                         radius = 0.5, taxonKey = 7707728, 
                         limit = 10000, backbone_list = NA)